# swagger/docs.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "CineReviews",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000/",
      "description": "Local Server"
    }
  ],
  "paths": {
    "/movies": {
      "get": {
        "description": "Get movies",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "$ref": "#/components/schemas/page"
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "$ref": "#/components/schemas/size"
          },
          {
            "name": "sortParam",
            "in": "query",
            "required": false,
            "$ref": "#/components/schemas/sortParam"
          },
          {
            "name": "sortDirection",
            "in": "query",
            "required": false,
            "$ref": "#/components/schemas/sortDirection"
          }
        ],
        "responses": {
          "200": {
            "description": "Movies were obtained",
            "content": {
              "text/html": {
                "schema": {
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "text/html": {
                "schema": {
                }
              }
            }
          }
        }
      }
    },
...
    "/movies/genres": {
      "get": {
        "description": "Get movies filtered by genre",
        "parameters": [
          {
            "name": "genreId",
            "in": "query",
            "required": false,
            "$ref": "#/components/schemas/Genre/properties/id"
          }
        ],
        "responses": {
          "200": {
            "description": "Movies were obtained",
            "content": {
              "text/html": {
                "schema": {
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "text/html": {
                "schema": {
                }
              }
            }
          }
        }
      }
    },
...
    "/actors": {
      "get": {
        "description": "Get Actors",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "$ref": "#/components/schemas/page"
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "$ref": "#/components/schemas/size"
          },
          {
            "name": "sortParam",
            "in": "query",
            "required": false,
            "$ref": "#/components/schemas/sortParam"
          },
          {
            "name": "sortDirection",
            "in": "query",
            "required": false,
            "$ref": "#/components/schemas/sortDirection"
          }
        ],
        "responses": {
          "200": {
            "description": "Actors were obtained",
            "content": {
              "text/html": {
                "schema": {
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "text/html": {
                "schema": {
                }
              }
            }
          }
        }
      }
    },
...
    "/crew": {
      "get": {
        "description": "get list of the crew",
        "responses": {
          "200": {
            "description": "Crews were obtained",
            "content": {
              "text/html": {
                "schema": {
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "text/html": {
                "schema": {
                }
              }
            }
          }
        }
      }
    },
    "/genres": {
      "get": {
        "description": "Retrieve a list of all genres.",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0
            },
            "description": "The page number for pagination."
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "The number of items per page."
          }
        ],
        "responses": {
          "200": {
            "description": "A list of genres was retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Genre"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 500
                    },
                    "message": {
                      "type": "string",
                      "example": "Internal Server Error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/languages": {
      "get": {
        "description": "Retrieve a list of all languages.",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0
            },
            "description": "The page number for pagination."
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "The number of items per page."
          }
        ],
        "responses": {
          "200": {
            "description": "A list of languages was retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Language"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 500
                    },
                    "message": {
                      "type": "string",
                      "example": "Internal Server Error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    },"/themes": {
      "get": {
        "description": "Retrieve a list of all themes.",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0
            },
            "description": "The page number for pagination."
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "The number of items per page."
          }
        ],
        "responses": {
          "200": {
            "description": "A list of themes was retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Theme"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 500
                    },
                    "message": {
                      "type": "string",
                      "example": "Internal Server Error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "page": {
        "type": "integer",
        "description": "number of the page if Pagination is used"
      },
      "size": {
        "type": "integer",
        "description": "size of the page if Pagination is used"
      },
      "sortParam": {
        "type": "string",
        "description": "parameter used to filter the data"
      },
      "sortDirection": {
        "type": "string",
        "enum": [
          "ASC",
          "DESC"
        ],
        "description": "parameter used to chose the filtering order"
      },
...
      "Genre": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Unique identifier for the genre",
            "example": 1
          },
          "genre": {
            "type": "string",
            "description": "Genre name",
            "example": "Comedy"
          }
        }
      },
      "Theme": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Unique identifier for the theme",
            "example": 1
          },
          "theme": {
            "type": "string",
            "description": "short description of the theme",
            "example": "Humanity and the world around us"
          }
        }
      },
      "Language": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Unique identifier for the language",
            "example": 1
          },
          "theme": {
            "type": "string",
            "description": "language specification",
            "example": "English"
          }
        }
      }
    }
  }
}